+++ core/r4300/bc.cpp
@@ -49,7 +49,7 @@ void BC1F()
 void BC1F_OUT()
 {
     if (check_cop1_unusable()) return;
-    jump_target = (long)PC->f.i.immediate;
+    jump_target = (int32_t)PC->f.i.immediate;
     PC++;
     delay_slot = 1;
     PC->ops();
@@ -63,7 +63,7 @@ void BC1F_OUT()
 
 void BC1F_IDLE()
 {
-    long skip;
+    int32_t skip;
     if ((FCR31 & 0x800000) == 0)
     {
         update_count();
@@ -92,7 +92,7 @@ void BC1T()
 void BC1T_OUT()
 {
     if (check_cop1_unusable()) return;
-    jump_target = (long)PC->f.i.immediate;
+    jump_target = (int32_t)PC->f.i.immediate;
     PC++;
     delay_slot = 1;
     PC->ops();
@@ -106,7 +106,7 @@ void BC1T_OUT()
 
 void BC1T_IDLE()
 {
-    long skip;
+    int32_t skip;
     if ((FCR31 & 0x800000) != 0)
     {
         update_count();
@@ -142,7 +142,7 @@ void BC1FL_OUT()
     if (check_cop1_unusable()) return;
     if ((FCR31 & 0x800000) == 0)
     {
-        jump_target = (long)PC->f.i.immediate;
+        jump_target = (int32_t)PC->f.i.immediate;
         PC++;
         delay_slot = 1;
         PC->ops();
@@ -159,7 +159,7 @@ void BC1FL_OUT()
 
 void BC1FL_IDLE()
 {
-    long skip;
+    int32_t skip;
     if ((FCR31 & 0x800000) == 0)
     {
         update_count();
@@ -195,7 +195,7 @@ void BC1TL_OUT()
     if (check_cop1_unusable()) return;
     if ((FCR31 & 0x800000) != 0)
     {
-        jump_target = (long)PC->f.i.immediate;
+        jump_target = (int32_t)PC->f.i.immediate;
         PC++;
         delay_slot = 1;
         PC->ops();
@@ -212,7 +212,7 @@ void BC1TL_OUT()
 
 void BC1TL_IDLE()
 {
-    long skip;
+    int32_t skip;
     if ((FCR31 & 0x800000) != 0)
     {
         update_count();
